% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mani_carry.R
\name{cl_wday_date}
\alias{cl_wday_date}
\title{Set Date by weekday}
\usage{
cl_wday_date(year, month, day, weekday = 3)
}
\value{
The closest weekday's date to the given date.
}
\description{
Set Date by weekday
}
