% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_dates.R
\name{cl_wed_date}
\alias{cl_wed_date}
\title{Set Date by weekday}
\usage{
cl_wed_date(year, month, day)
}
\value{
The date of the closest Wednesday to the given date.
}
\description{
Set Date by weekday
}
