% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nye2normal.R
\name{compare_nye2normal}
\alias{compare_nye2normal}
\title{Get New Year's Eve & normal Wednesdays stats of a given year}
\usage{
compare_nye2normal(carry, year)
}
\description{
Get New Year's Eve & normal Wednesdays stats of a given year
}
