% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nye2normal.R
\name{compare_nye2wdays}
\alias{compare_nye2wdays}
\title{Get New Year's Eve & normal Wednesdays stats of a given year}
\usage{
compare_nye2wdays(carry, year, groupby = "station")
}
\description{
Three columns, \code{date} (class: Date), \code{num_in},
and \code{num_out} must be present in \code{carry}.
}
