% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mani_carry.R
\name{filter_stations}
\alias{filter_stations}
\title{Given the data \code{carry}, retrieve the filtered data.}
\usage{
filter_stations(carry, sta = "台北", years = 2005, months = 1:12)
}
\description{
Given the data \code{carry}, retrieve the filtered data.
}
