% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_dates.R
\name{is_working_day_atom}
\alias{is_working_day_atom}
\title{Is the date given 'normal' (Not holidays, not weekends)}
\usage{
is_working_day_atom(year, month, day)
}
\description{
Is the date given 'normal' (Not holidays, not weekends)
}
