% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_nye2normal.R
\name{to_sf}
\alias{to_sf}
\title{Convert data frame to sf spatial data frame}
\usage{
to_sf(carry_nye_normal, stations)
}
\description{
Merge data returned from \code{compare_nye2wdays} with
sf datafrme \code{stations}.
}
